modularize all components
