+++ src/components/Header.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Header({ incomplete }) {
+  return (
+    <header>
+        <h1>To Do List: {incomplete} items pending</h1>
+      </header>
+  )
+}
+
+export default Header;
