+++ src/components/Item.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function Item({item, toggleComplete}) {
+  return (
+    <>
+        <p>{item.text}</p>
+        <p><small>Assigned to: {item.assignee}</small></p>
+        <p><small>Difficulty: {item.difficulty}</small></p>
+        <div onClick={() => toggleComplete(item.id)}>Complete: {item.complete.toString()}</div>
+        <hr />
+    </>
+  )
+}
+
+export default Item;
