+++ src/components/todo.js
@@ -1,23 +1,12 @@
 import React, { useEffect, useState } from 'react';
-import useForm from '../hooks/form.js';
-
-import { v4 as uuid } from 'uuid';
+import Header from './Header';
+import ItemForm from './ItemForm';
+import ItemList from './ItemList';
 
 const ToDo = () => {
 
-  const [defaultValues] = useState({
-    difficulty: 4,
-  });
   const [list, setList] = useState([]);
   const [incomplete, setIncomplete] = useState([]);
-  const { handleChange, handleSubmit } = useForm(addItem, defaultValues);
-
-  function addItem(item) {
-    item.id = uuid();
-    item.complete = false;
-    console.log(item);
-    setList([...list, item]);
-  }
 
   function deleteItem(id) {
     const items = list.filter( item => item.id !== id );
@@ -34,7 +23,6 @@ const ToDo = () => {
     });
 
     setList(items);
-
   }
 
   useEffect(() => {
@@ -45,43 +33,9 @@ const ToDo = () => {
 
   return (
     <>
-      <header>
-        <h1>To Do List: {incomplete} items pending</h1>
-      </header>
-
-      <form onSubmit={handleSubmit}>
-
-        <h2>Add To Do Item</h2>
-
-        <label>
-          <span>To Do Item</span>
-          <input onChange={handleChange} name="text" type="text" placeholder="Item Details" />
-        </label>
-
-        <label>
-          <span>Assigned To</span>
-          <input onChange={handleChange} name="assignee" type="text" placeholder="Assignee Name" />
-        </label>
-
-        <label>
-          <span>Difficulty</span>
-          <input onChange={handleChange} defaultValue={defaultValues.difficulty} type="range" min={1} max={5} name="difficulty" />
-        </label>
-
-        <label>
-          <button type="submit">Add Item</button>
-        </label>
-      </form>
-
-      {list.map(item => (
-        <div key={item.id}>
-          <p>{item.text}</p>
-          <p><small>Assigned to: {item.assignee}</small></p>
-          <p><small>Difficulty: {item.difficulty}</small></p>
-          <div onClick={() => toggleComplete(item.id)}>Complete: {item.complete.toString()}</div>
-          <hr />
-        </div>
-      ))}
+      <Header incomplete={incomplete}/>
+      <ItemForm list={list} setList={setList}/>
+      <ItemList list={list} toggleComplete={toggleComplete} />
 
     </>
   );
